requirements.txt
botcity-framework-web>=0.8.0,<1.0
botcity-maestro-sdk>=0.3.3,<1.0
botcity-twilio-whatsapp-plugin==0.1.3
